Version 0.0.4
Created an authorization middleware for API routes so no information can
be retrieved or posted without an access token.
Fixed a bug in the template order which resulted in information title
being behind the clock.
Removed an unnescessary directive.
Created instructions on how to disable authentication if it is not
needed.
Updated the to-do list.